remove CredentialsBuilder
decided to move all the interactions with the config hash into
SoapRequestBuilder. Originally my thought was to inject CredentialsBuilder
into SoapRequestBuilder so SoapRequestBuilder did not have to know about both
endpoints and credentials. However, that means we have 2 objects that have
to know how the config hash works. Right now it seems easier just to have
one.

If we revisit this it might make sense to have an EndpointBuilder and a
CredentialsBuilder. An Endpoint object might prove useful; it could possibly
provide some configuration to the curl command in the execute_request.