add conversion for meeting patterns' days
the data in the xml is all days of the week, with the active ones identified
by a day abbreviation. Select the days that have an abbreviation and map
then into an array for Day objects.

note: I am using the xml tag to do the mapping, since there is no example in
the data of what an active sunday's abbreviation would be