add CleAttributes
CleAttributes contains CleAttribute objects and knows how to merge
collections of them.

Right now the merge process returns a new instance of CleAttributes.
This is an experiment with immutability. Since in the end we will be
doing something along the lines of many_course_rows.inject(:merge), it
seems like a good idea not to mutate the rows as we go along. If this
ends up being a source of bugs or a memory issue/non-performant we can
change this.