update row method list
we were fetching the list of xml data translation methods from the
Row::ROW_ATTRIBUTES hash but this was ignoring the method for the meeting
pattern days which was defined directly in the class, not the ROW_ATTRIBUTES
hash. (The days method deals with values from multiple XML tags so there is
no simple mapping.)

Get the list of methods from the class's instance methods (and removing
the generic methods provided by Object) instead.

Should probably name namespace the course methods (e.g. course_aspect__course_id)
so we can switch row_values_for(nil) to row_values_for('course_aspect') in
the course_aspect method. That way we can add other methods to Row and they
won't get pulled in when we are building a course aspect. (That was not a danger
before because we were using the configuration hash)