allow Resource classes to override their json 'type'
This is a workaround for a mismatch between my strategy for parsing
and aggregating this xml and the expected json output. Each row in the
xml is one branch of the course object tree, so I call what gets built
from that partial information a CourseAspect. Once we merge all these
branches together we have a course, and that is achieved by calling
.merge on all the aspects. Because that is a collection operation, the
returned object is still a CourseAspect, but the json calls it a
'course'.

We could feed this merged CourseAspect into a new Course object, but
that seems redundant; Course would have the exact same behavior as
CourseAspect. I might go that way but for now it seems simpler to let
CourseAspect have a type of 'course' for the purposes of json (or other)
representations.