add a daemon task for the download
hooks into a rake task.

rufus doesn't want to find the rake task even though the daemon-kit
config should be loading the files in tasks/ so call the task with
rake -f

require the class_json_export code. Need to look at the daemon
configuration options. There has got to be a better way than to
require it twice