fix values in resource and resource colllection tests
Resource assumes it gets initialized with arguments that respond to
json_tree, so we need to update the test setup to use the Value::String,
Value::Integer, etc. objects.

A better solution might be to have Resource do this casting (if necessary)
on initialization, but for now lets get the test to stop breaking