wrap rake tasks in something that will raise errors
rake is sends any errors in the tasks to stderr and not raising
them. They need to be raised so that daemon kit's error logging
and other hooks can be triggered.

add RakeRunner, which runs the commands through Open3.popen3 so we can
inspect stderr and raise an exception if there has been any errors