add integration tests for missing xml resource data
when the xml has no information for a subresource that node should
be missing from the json. This was previously tested at a unit level in
resource_spec.rb. This commit adds xml file to json file integration
tests for this behavior

no_grading_basis.xml and no_instruction_mode.xml are subsets of our
reference.xml with data removed. The tests make sure the key for
the missing resource is missing from the parsed json