Add object to handle configuration of FileNames
FileNames is being created with a bunch of different options. A hash,
some strings, etc. @dlaggero and I decided to extract that out to a
config object.

A configuration option we're introducing here is the 'service'. Our
first task for this story is to be able to extract 2 xml files from 2
different services (Courses and Classes). FileNames will have to know
about service, so we've added to the config object.