Begin refactor of extraction to workflow of steps
This refactor starts the process of moving ClassJsonExport from an
object that knows all the steps of the workflow to an Orchestrator of
several smaller workflow-step objects.

The plan here is to start the workflow off with a single 'step' and then
have each step tell the orchestrator to call the subsequent step. So our
first step, BuildSources, calls `run_step` and passes in GetData.

This approach is different from having the Orchestrator know a static
collection of steps, and then cycle through them in order. Our approach
is a little trickier, but offers more flexibility. For example, if a
step fails, it can call `run_step` with the Terminate step. Or,
possibly, one step might have to dynamically determine which step to
call next.