add ParseData step
ParseData#run receives the data sources and the orchestrator. It
finds a parser for each data source and uses it to parse the data
source's data. It then passes the collection of parsed data on to
MergeData

This commit also adds methods on ParseData's collaborators.
DataSource#service_name is used to find the parser and ClassJson.parse
will parse the xml. We still need a CourseJson parser and to fill in
DataSource#data.