add reference_combined.json
reference_combined.json contains the json we wnat after the course
and class xml files have been merged. It is the reference.json from
the first iteration with the course title replaced with the value from
the course xml and it adds a frequency (e.g. 'Every Fall and Spring')
from the course.xml. All other values are teh smae as before and come
from the class xml

leaving reference.json in place for now. We are in the middle of a
refactoring so want to leave it for existing tests.