Refactor the merge_data test
Now that the data is sorted in a previous step, course_id is not
significant and whether the data is from the course source or the class
source is not important.  Not positive this is correct, but want to get
this into github so we work can continue while I am out.