Give QueryResults a declarative as_json method
QueryResults only has one return value we need, the json output. So
instead of creating an instance and calling `to_json` we're now just
building it `as_json`. Implementation is the same.

Internally we renamed the instance's `to_json` method to `as_json` as
well, since ruby objects already implement a `to_json` method that we
were stomping on.