add tests for DataSource#data
DataSource#data builds the data request for the DataSource's service
and returns the results from querying the service with the institution,
campus and term from the query config.

Because the services can return data back in serveral responses, all
responsed are gathered together and wrapped in a base tag so the responses
look like a single xml document to parsers.