change base tag in DataSource#data response
DataSource#data was wraps all responses from the services in a tag
so multi-part responses act as a single document. The name of that
tag was hardcoded to 'class_service_data' which is no longer
appropriate now that we are also getting data from the course service.
Change the tag so it contains the name of the service e.g
'course_service_data' for courses 'class_service_data' for classes