change Resource#new to take a hash of attributes
the old method tied the initialization of resource classes (CourseAspect
Section, etc.) to the order that the attributes are defined in the
class's attrbiutes and child_collections arrays. A side effect of this
was that the CourseAspectBuilder had to re-arrange arguments and/or
the order methods in ROW_ATTRIBUTES had to match the order of these
arrays in order for the objects to get properly built.

This refactoring changes Resource#initialize so it takes a hash of
attributes. This breaks the coupling to ordering in Row and CourseAspectBuilder.
We need to make sure the last 2 parts of the method name key in the
ROW_ATTRIBUTES hash match the class and attribute names (e.g. a
mapping for Section#awesomeness_level needs to end in
'section__awesomeness_level')

Unlike typical attribute hash processing, we need to add an extra check
to see if the 'attribute' corresponds to one of the child collections and,
if so wrap the value in a ResourceCollection so it can be merged, converted
to json, etc.