Expose campus and term from DataSource
The ParseData step is going to need to know the campus and term.
Currently it is getting them by parsing a file name, but that code is
going to go away. So we've chosen to allow a DataSource instance provide
the campus and term, since DataSource instances are being passed to the
ParseData code.