Use data source while parsing, remove file logic
A few changes to the ClassJson implementation here:

First, we're using the provided data source to get the xml_string,
campus and term. We're no longer getting these attributes from a file or
the file's name.

Second, we're no longer writing a file. The whole `to_file` method is
gone because that is now handled at the ClassJsonExport level.

Third, we're no longer doing the merging inside the `courses` method.
That has moved to the MergeData class. All ClassJson needs to do is
return a bunch of unmerged, unsorted CourseAspects.

Finally, the test of ClassJson only looks at the return value of
`parse`, as that's the only method that clients of this code use. We
talked about mocking out all of the collaborators that are used to
generate the returned json, but decided to go to a much simpler
behavioral test that compares the returned json to a string.

All of these changes follow on from the refactorings we've done in
previous commits. Much of this was vestigial code that we intended to
remove once the refactoring was done.