Replace Class|CourseJson with Class|CourseParser
Class and CourseJson were misnamed as they no longer accepted or
returned json. What they do is parse xml into CourseAspects. I've
changed their name to make it a bit better.

Along with this, I extracted all the shared behavior into a super-class,
abstract_parser. This is an intermediate step along a larger
refactoring, so the goofy name is not a big concern.