Greatly reduce amount of requiring
The typical practice we've been following is to have one file do a
require of all the files in a subdirectory. For example, the file
lib/query.rb would require all the files in lib/query/

I've followed that pattern here, with a few exceptions where a file has
to be loaded before all others.

For now I've removed all requires in tests. We've been finding that
tests can pass because they require files that aren't required when the
daemon runs. I like the idea of tests being explicit about what files
they require, but it also has pitfalls.