Extract shared Row and Rows logic to superclasses
Consolidating the knowledge shared between ClassRow/CourseRow and
ClassRows/CourseRows.

Because of the way the attribute methods are being metaprogrammed in,
each subclass has to have that loop through ROW_ATTRIBUTES. I'm sure
there's a way to fix that, but I don't know off the top of my head.

And the metaprogramming aspect seems to be untested. When I broke it,
all the tests still passed. So before that bit of code can be
refactored, tests should be added.