Fixes the merge for course aspect
Now that we have multiple sources for our course aspect, we need to
merge the attributes of the course aspect, not just the subcollections
that the course aspect has.  If there are differences between the
original attribute and the new attribute, the original attribute has
priority unless it is blank. If we want to add new attributes, the
attribute should not already be mapped in the existing row class
(ClassRow or CourseRow).