move course attributes to CourseRow
the course service returns all active courses for the term, the
class service just returns those that are being offered for the term.

the consumer of our processed json file needs to know about all courses
whether they are offered or not. our old implementation which got course
level information from the class service resulted in missing information
when there was no offering for that term. Using the course service's data
fixes this.