change how CourseAspect determine's attribute to merge
the old implementation used `#blank?` which comes from active support.
We are only using the inflection methods from active support so `blank?`
should not be available (even though it is when you run bundle exec rake).
Change our implementation to no longer rely on this method.

Also we changed the test setup in course_aspect_spec to wrap values in
the XmlParser::Value classes to better reflect production behavior