get course_attribute from the course xml
Course Attributes need to be available in a full course catalog. The prior
practice of getting the data from the class xml means that courses that do
not have a class for a term will not have any course attribute data.

Also update the reference_courses.xml. The download that the reference
data was extracted from contained extra attributes, e.g. DELM delivery method,
that we do not care about, and this was breaking the integration test.