change RakeRunner#run to work with Rake#sh
By default Rake#sh sends the command to stderr before running it
(see Rake#rake_output_message). This looks like an error to
RakeRunner, so for rake tasks that used sh, RakeRunner was raising
exceptions even though the command was succeeding.

Change the implementation of RakeRunner#run to check for the return
value of the command and the contents of stderr. This prevents
rake_output_message's use of stderr from causing an exception to be raised.