add Grouping
Grouping does the work of the SortData step. It receives collections
of items and returns new collections of those items, grouped by
the supplied attribute.

For example, in our SortData step we start with 2 collections of
CourseAspects from the ClassService and CourseService and what we want
to pass to the MergeData step is a collection where all the
CourseAspects for course_id 1 are gathered together and the
CourseAspects for course_id 2 are gathered together, etc.