use Grouping in SortData
The tests for SortData were a little tricky because we could not make
fine grained assertions about the results that were getting passed on
to MergeData and we were working around this by specifying an additional
unneeded sort in our implementation.

The previous commit extracted this sorting behavior into a Grouping class
(grouping is closer to what we are doing in this step) that can be tested
in a more fine grained way. This commit re-implements SortData using
the new class