enforce a maximum number for attempts for GetQueryResults
we are running into an issue where the query instance we are getting
handed is not the query instance results are getting sent to. As
things currently stand this means we will keep polling for data
forever.

While the integration broker issue needs to get fixed, the current
behavior of the code is not good and we need to exit at some point.

Add a PollingConfiguration object that will supply the maximum attempts.
If we have polled too many times, raise an error and stop.

The polling code is asking for a refactor