Take 2 - smarter merge
Merging everything all at once took way way too long. Logs show UMNTC
summer 2016 went from 12-15 minutes to several house

Whereas before sorting the whole collection took a long time, but
merging the subsets was slow but not too bad.

In both cases the challenge seems to be dealing with a large collection.

THe new approach goes through the parsed course aspects once but instead
of trying to merge the new course info onto one big CampusTermCourse
object we merge it onto an exisiting course (if there is one, otherwise
just add it to the collection). The hope is that we can still avoid the
sorting step while preserving the smaller sized merges.