Update Ruby
The Courses API is moving to a new server that runs Ruby 2.7. This repo
needs to be running on the same server as the Courses API, so updating
this repo.

Throwing bundler-audit in there so we can easily check for gems with
CVE. I've addressed all issues that it found.