Update Capistrano and its dependencies
Capistrano hasn't been updated in this application since it was first
added 6 years ago. Updating it as there are metrics capabilities in the
old version that we don't want to use.

The gem has to be locked to a specific version that is referenced in
config/deploy. Running the tests, all pass.