Upgrade to Ruby v3.1.6
In preparation for containerization, this PR upgrades the Ruby version to v3.1.6.

Because the project uses the `daemon-kit` gem, and that gem has not been updated to support Ruby v3.2+, we are stuck using Ruby v3.1 for the time being. That said, I plan to remove `daemon-kit` completely in a future PR. This Ruby version upgrade is just a stepping stone to future upgrades.

This PR also upgrades the `activesupport` and `i18n` gems address advisories logged by `bundler-audit`. Additionally, the `.gitignore` file was updated to include the `config/credentials.yml` file which is created as part of `script/setup`.